Move the stored identity type to the model layer
